third_party/WebKit/LayoutTests/crypto/subtle/ecdh/deriveBits-expected.txt

Tests ECDH's deriveBits() using P-521

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Importing the private key...

Importing the public key...

Deriving 521 bits...

PASS: Derived Bytes should be [0117d54d84379d0fd385be068455a77a5366ab534ff172ab0a121f37d180dccd19607abb0c41cb9f6f12b01303ac4a69dc2d1d05180181fd496d9769b46bffec3400] and was
PASS successfullyParsed is true

TEST COMPLETE
